refactor(agents): extract activity seam from agents.ts
Move ActivityEntry, appendActivity, getActivity, saveSessionId, getSessionId,
getLatestSessionIdSync, and getLatestSessionId to src/lib/agents/activity.ts;
re-export from agents.ts.

Behavior-preserving move only; no logic changes.
